pyproject.toml
[tool.poetry]
name = "nonebot-plugin-qrcode"
version = "0.0.1"
description = "qq聊天二维码插件"
authors = ["kexue <x@kexue.io.com>"]
license = "MIT"
homepage = "https://github.com/kexue-z/nonebot-plugin-qrcode"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7.3,<4.0.0"
pillow = "^8.3.1"
nonebot2= "^2.0.0.beta.1"
nonebot-adapter-onebot = "^2.0.0.beta.1"
pyzbar-x = "^0.2.1"
httpx = ">=0.18.0,<1.0.0"
qrcode = "^7.0"


[tool.poetry.dev-dependencies]
black = {version = "^21.10b0", allow-prereleases = true}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
